chore(macos): remove dead symbols left by the settings retirement
Remove unused native settings and onboarding chat code, obsolete presentation helpers and tests, and the unsubscribed permission poller. Preserve explicit permission refresh, live exec-approval readers and migration/socket behavior, and Dashboard onboarding handoff coverage. Retain the SwiftUI app delegate through the existing narrowly scoped Periphery convention; let the Advanced disclosure own its expansion state.

Replace obsolete native inventory assertions with surviving Connection and consent copy. Regenerate only the source inventory; leave Localizable.xcstrings unchanged. Periphery 3.8.0 strict scan is clean with the local native SwiftPM build override. App/test builds, 38 pure Swift tests, 35 i18n tests, Swift lint/format, i18n verification, check-changed, and fresh autoreview passed.
