--- apps/macos/Sources/OpenClaw/DurationFormattingSupport.swift
@@ -1,15 +1,0 @@
-import Foundation
-
-enum DurationFormattingSupport {
-    static func conciseDuration(ms: Int) -> String {
-        if ms < 1000 { return "\(ms)ms" }
-        let s = Double(ms) / 1000.0
-        if s < 60 { return "\(Int(round(s)))s" }
-        let m = s / 60.0
-        if m < 60 { return "\(Int(round(m)))m" }
-        let h = m / 60.0
-        if h < 48 { return "\(Int(round(h)))h" }
-        let d = h / 24.0
-        return "\(Int(round(d)))d"
-    }
-}
