+++ apps/macos/Sources/OpenClaw/PermissionManager.swift
@@ -3,7 +3,6 @@ import ApplicationServices
 import AVFoundation
 import CoreLocation
 import Foundation
-import Observation
 import OpenClawIPC
 import PeekabooAutomationKit
 import Speech
@@ -418,71 +417,21 @@ final class LocationPermissionRequester: NSObject, CLLocationManagerDelegate {
 }
 
 @MainActor
-@Observable
 final class PermissionMonitor {
     static let shared = PermissionMonitor()
 
-    private(set) var status: [Capability: CapabilityAuthorizationStatus] = [:]
-
-    private var monitorTimer: Timer?
+    private var status: [Capability: CapabilityAuthorizationStatus] = [:]
     private var isChecking = false
-    private var registrations = 0
-    private var lastCheck: Date?
-    private let minimumCheckInterval: TimeInterval = 0.5
-
-    func register() {
-        self.registrations += 1
-        if self.registrations == 1 {
-            self.startMonitoring()
-        }
-    }
-
-    func unregister() {
-        guard self.registrations > 0 else { return }
-        self.registrations -= 1
-        if self.registrations == 0 {
-            self.stopMonitoring()
-        }
-    }
 
     func refreshNow() async {
-        await self.checkStatus(force: true)
-    }
-
-    private func startMonitoring() {
-        Task { await self.checkStatus(force: true) }
-
-        if ProcessInfo.processInfo.isRunningTests { return }
-        self.monitorTimer = Timer.scheduledTimer(withTimeInterval: 1.0, repeats: true) { [weak self] _ in
-            guard let self else { return }
-            Task { @MainActor in
-                await self.checkStatus(force: false)
-            }
-        }
-    }
-
-    private func stopMonitoring() {
-        self.monitorTimer?.invalidate()
-        self.monitorTimer = nil
-        self.lastCheck = nil
-    }
-
-    private func checkStatus(force: Bool) async {
         if self.isChecking { return }
-        let now = Date()
-        if !force, let lastCheck, now.timeIntervalSince(lastCheck) < self.minimumCheckInterval {
-            return
-        }
-
         self.isChecking = true
+        defer { self.isChecking = false }
 
         let latest = await PermissionManager.authorizationStatus()
         if latest != self.status {
             self.status = latest
             NotificationCenter.default.post(name: .openclawPermissionsChanged, object: nil)
         }
-        self.lastCheck = Date()
-
-        self.isChecking = false
     }
 }
