+++ apps/macos/Tests/OpenClawIPCTests/StatusMenuReconcileTests.swift
@@ -154,9 +154,6 @@ struct StatusMenuReconcileTests {
             sessionId: nil,
             thinkingLevel: nil,
             verboseLevel: nil,
-            systemSent: false,
-            abortedLastRun: false,
-            tokens: SessionTokenStats(input: 0, output: 0, total: 0, contextTokens: 200_000),
-            model: nil)
+            tokens: SessionTokenStats(total: 0, contextTokens: 200_000))
     }
 }
